Rework tests and add tests for the new way to call components
Separate tests between existing syntax and the new syntax. Also test if the exception is thrown when the syntax is not respected.
