+++ tests/views/invalid-type-test.twig
@@ -1,0 +1,5 @@
+{% extends 'layout' %}
+
+{% block content %}
+    {% livewire 63 %}
+{% endblock %}
