+++ tests/views/layout.twig
@@ -1,9 +1,5 @@
 {{ livewireStyles() }}
 
-{% livewire counter %}
-
-{% livewire counter with {'count': 3} %}
-
-{% livewire counter with {'count': 3, 'title': 'Lorem ipsum!'} %}
+{% block content %}{% endblock %}
 
 {{ livewireScripts() }}
